requirements.txt
Django>=5.0.2
django-environ==0.11.2
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.54.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
python-dotenv==1.0.0
channels>=4.0.0
whitenoise==6.5.0 